Move param and ident
