--- src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ident.rs
@@ -1,9 +1,0 @@
-use swc_ecma_ast::Ident;
-
-use super::GetName;
-
-impl GetName for Ident {
-    fn get_name(&self) -> &str {
-        self.sym.chars().as_str()
-    }
-}
