+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/traits.rs
@@ -1,8 +1,14 @@
 use swc_common::Span;
-use swc_ecma_ast::TsType;
+use swc_ecma_ast::{Ident, TsType};
 
 pub trait GetName {
     fn get_name(&self) -> &str;
+}
+
+impl GetName for Ident {
+    fn get_name(&self) -> &str {
+        self.sym.chars().as_str()
+    }
 }
 
 pub trait GetSourceInfo {
